refactor: JanggiChessPiece의 움직임 추상화 제거 및 JanggiPiece 인터페이스 삭제
