--- src/main/java/domain/janggiPiece/JumpJanggiPiece.java
@@ -1,29 +1,0 @@
-package domain.janggiPiece;
-
-import domain.direction.JanggiDirections;
-import domain.path.Path;
-import domain.position.JanggiPosition;
-import domain.type.JanggiTeam;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public abstract class JumpJanggiPiece extends JanggiChessPiece {
-    private final List<JanggiDirections> directions;
-
-    protected JumpJanggiPiece(JanggiTeam team, List<JanggiDirections> directions) {
-        super(team);
-        this.directions = directions;
-    }
-
-    @Override
-    public final List<Path> getCoordinatePaths(JanggiPosition startPosition) {
-        List<Path> result = new ArrayList<>();
-        for (JanggiDirections direction : directions) {
-            if (direction.canApplyFrom(startPosition)) {
-                result.add(direction.getPathFrom(startPosition));
-            }
-        }
-        return result;
-    }
-}
