+++ src/main/java/domain/position/generator/JanggiPiecePositionsGenerator.java
@@ -1,10 +1,10 @@
 package domain.position.generator;
 
-import domain.janggiPiece.JanggiPiece;
+import domain.janggiPiece.JanggiChessPiece;
 import domain.position.JanggiPosition;
 
 import java.util.Map;
 
 public interface JanggiPiecePositionsGenerator {
-    Map<JanggiPosition, JanggiPiece> generate();
+    Map<JanggiPosition, JanggiChessPiece> generate();
 }
