+++ src/test/java/domain/position/generator/EmptyPositionsGenerator.java
@@ -1,6 +1,6 @@
 package domain.position.generator;
 
-import domain.janggiPiece.JanggiPiece;
+import domain.janggiPiece.JanggiChessPiece;
 import domain.position.JanggiPosition;
 
 import java.util.Map;
@@ -8,7 +8,7 @@ import java.util.Map;
 public class EmptyPositionsGenerator implements JanggiPiecePositionsGenerator {
 
     @Override
-    public Map<JanggiPosition, JanggiPiece> generate() {
+    public Map<JanggiPosition, JanggiChessPiece> generate() {
         return Map.of();
     }
 }
